+++ Tests/ConsoleTests/CriticalSectionTests.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace ConsoleTests
+{
+    static class CriticalSectionTests
+    {
+        public static void Start()
+        {
+
+        } 
+    }
+}
